rip out species computing code from project details
